feat(nomina): la ficha guarda fondo de cesantias y caja de compensacion; catalogo de cajas
No habia donde elegir ninguno de los dos. El fondo tenia columna heredada del SOLIDO
(SeveranceFundId, decimal(6,0)) que nada escribia ni leia; la caja era un entero sin
tabla. Ahora: FamilyCompensationFund (PAY_FamilyCompensationFunds) con CRUD, endpoints
/api/payroll/family-compensation-funds y pantalla Nomina > Cajas de compensacion,
calcados del catalogo de cesantias; RegisterEmployee/UpdateEmployee reciben
severanceProviderPublicId y familyCompensationFundPublicId, el detalle los devuelve con
nombre, y la ficha tiene los dos desplegables. Migracion en par (SqlServer y PostgreSql):
SeveranceFundId pasa a entero (sin perdida: siempre valio 0) y nace la tabla de cajas.
Ninguno de los dos cambia la liquidacion mensual. e2e: PUT con fondo y caja -> el detalle
los devuelve; suites Architecture y Application en verde.

+++ src/Core/IngenIA365ERP.Domain/Entities/Payroll/Employee.cs
@@ -91,7 +91,8 @@ public class Employee : AuditableEntity
     public int PensionFundId { get; set; }
     public int WorkRiskId { get; set; }
     public int WorkRiskRateId { get; set; }
-    public decimal SeveranceFundId { get; set; }
+    /// <summary>Fila de <c>PAY_SeveranceProviders</c>. Era decimal(6,0) heredado del SOLIDO; entero desde el 2026-09-12.</summary>
+    public int SeveranceFundId { get; set; }
     public int SenaId { get; set; }
     public int IcbfId { get; set; }
     public int FamilySubsidyId { get; set; }
